rename according to python cconventions
