+++ src/nexgen/beamlines/I19_2_gda_nxs.py
@@ -22,9 +22,9 @@ from ..nxs_utils import (
     Source,
     TristanDetector,
 )
-from ..nxs_utils.Detector import DetectorType, UnknownDetectorTypeError
-from ..nxs_utils.ScanUtils import calculate_scan_points
-from ..nxs_write.NXmxWriter import EventNXmxFileWriter, NXmxFileWriter
+from ..nxs_utils.detector import DetectorType, UnknownDetectorTypeError
+from ..nxs_utils.scan_utils import calculate_scan_points
+from ..nxs_write.nx_mx_writer import EventNXmxFileWriter, NXmxFileWriter
 from ..utils import get_iso_timestamp, get_nexus_filename
 from .beamline_utils import BeamlineAxes, collection_summary_log
 from .GDAtools.ExtendedRequest import (
