+++ src/nexgen/beamlines/I19_2_nxs.py
@@ -24,10 +24,10 @@ from ..nxs_utils import (
     Source,
     TristanDetector,
 )
-from ..nxs_utils.ScanUtils import calculate_scan_points, identify_osc_axis
-from ..nxs_write.NXmxWriter import EventNXmxFileWriter, NXmxFileWriter
-from ..tools.Metafile import DectrisMetafile
-from ..tools.MetaReader import define_vds_data_type, update_axes_from_meta
+from ..nxs_utils.scan_utils import calculate_scan_points, identify_osc_axis
+from ..nxs_write.nx_mx_writer import EventNXmxFileWriter, NXmxFileWriter
+from ..tools.metafile import DectrisMetafile
+from ..tools.meta_reader import define_vds_data_type, update_axes_from_meta
 from ..utils import find_in_dict, get_iso_timestamp, get_nexus_filename
 from .beamline_utils import collection_summary_log
