+++ src/nexgen/beamlines/SSX_Eiger_nxs.py
@@ -12,9 +12,9 @@ import h5py
 
 from .. import log
 from ..nxs_utils import Attenuator, Beam, Detector, EigerDetector, Goniometer, Source
-from ..nxs_write.NXmxWriter import NXmxFileWriter
-from ..tools.Metafile import DectrisMetafile
-from ..tools.MetaReader import define_vds_data_type, update_axes_from_meta
+from ..nxs_write.nx_mx_writer import NXmxFileWriter
+from ..tools.metafile import DectrisMetafile
+from ..tools.meta_reader import define_vds_data_type, update_axes_from_meta
 from ..utils import find_in_dict, get_iso_timestamp
 from .beamline_utils import PumpProbe, collection_summary_log
