+++ src/nexgen/beamlines/SSX_Tristan_nxs.py
@@ -10,7 +10,7 @@ from pathlib import Path
 
 from .. import log
 from ..nxs_utils import Attenuator, Beam, Detector, Goniometer, Source, TristanDetector
-from ..nxs_write.NXmxWriter import EventNXmxFileWriter
+from ..nxs_write.nx_mx_writer import EventNXmxFileWriter
 from ..utils import Point3D, find_in_dict, get_iso_timestamp
 from .beamline_utils import collection_summary_log
