+++ src/nexgen/command_line/cli_utils.py
@@ -24,8 +24,8 @@ from ..nxs_utils import (
     TransformationType,
     TristanDetector,
 )
-from ..nxs_utils.Detector import EIGER_CONST, UnknownDetectorTypeError
-from ..nxs_write.NXmxWriter import EventNXmxFileWriter, NXmxFileWriter
+from ..nxs_utils.detector import EIGER_CONST, UnknownDetectorTypeError
+from ..nxs_write.nx_mx_writer import EventNXmxFileWriter, NXmxFileWriter
 from ..utils import coerce_to_path, coord2mcstas, imgcif2mcstas
 
 logger = logging.getLogger("nexgen.cli_utils")
