+++ src/nexgen/nxs_utils/sample.py
