+++ src/nexgen/nxs_write/nx_class_writers.py
