+++ src/nexgen/tools/data_writer.py
@@ -14,7 +14,7 @@ import numpy as np
 from hdf5plugin import Bitshuffle
 from numpy.typing import ArrayLike
 
-from .Constants import (
+from .constants import (
     clock_freq,
     eiger_gap_size,
     eiger_mod_size,
