+++ tests/nxs_copy/test_copy_tools.py
@@ -12,7 +12,7 @@ from nexgen.nxs_copy.copy_utils import (
     identify_tristan_scan_axis,
     is_chipmap_in_tristan_nxs,
 )
-from nexgen.nxs_write.NXclassWriters import write_NXentry, write_NXnote
+from nexgen.nxs_write.nx_class_writers import write_NXentry, write_NXnote
 from nexgen.nxs_write.write_utils import create_attributes
 
 test_chipmap = {"01": (0, 0)}
