+++ tests/nxs_write/test_NXmxWriter.py
@@ -11,13 +11,13 @@ fake_gonio = Goniometer(
 )
 
 
-@patch("nexgen.nxs_write.NXmxWriter.write_NXsample")
-@patch("nexgen.nxs_write.NXmxWriter.write_NXsource")
-@patch("nexgen.nxs_write.NXmxWriter.write_NXdetector_module")
-@patch("nexgen.nxs_write.NXmxWriter.write_NXdetector")
-@patch("nexgen.nxs_write.NXmxWriter.write_NXinstrument")
-@patch("nexgen.nxs_write.NXmxWriter.write_NXdata")
-@patch("nexgen.nxs_write.NXmxWriter.write_NXentry")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXsample")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXsource")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXdetector_module")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXdetector")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXinstrument")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXdata")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXentry")
 def test_NXmxFileWriter_write(
     mock_NXentry,
     mock_NXdata,
@@ -39,21 +39,21 @@ def test_NXmxFileWriter_write(
     mock_NXsample.assert_called_once()
 
 
-@patch("nexgen.nxs_write.NXmxWriter.write_NXdatetime")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXdatetime")
 def test_NXmxFileWriter_updates_timestamps(mock_NXdatetime, dummy_NXmxWriter):
     fake_timestap = datetime.now()
     dummy_NXmxWriter.update_timestamps(fake_timestap, "start_time")
     mock_NXdatetime.assert_called_once()
 
 
-@patch("nexgen.nxs_write.NXmxWriter.write_NXnote")
+@patch("nexgen.nxs_write.nx_mx_writer.write_NXnote")
 def test_NXmxFileWriter_adds_note(mock_NXnote, dummy_NXmxWriter):
     fake_notes = {"foo": "bar"}
     dummy_NXmxWriter.add_NXnote(fake_notes)
     mock_NXnote.assert_called_once()
 
 
-@patch("nexgen.nxs_write.NXmxWriter.image_vds_writer")
+@patch("nexgen.nxs_write.nx_mx_writer.image_vds_writer")
 def test_NXmxFileWriter_write_vds(mock_vds_writer, dummy_NXmxWriter):
     dummy_NXmxWriter.goniometer = fake_gonio
     dummy_NXmxWriter.detector.detector_params.image_size = (100, 100)
